refactor: hoist RouteConfig + CreditsCallable to payments_py.x402.types
aaitor's follow-up #5: same pattern as the PaymentRequiredError
consolidation. Both the FastAPI middleware and the LangSmith Deployment
middleware had byte-identical RouteConfig dataclasses with the docstring
rationale "let extras diverge". The first divergent field would require
two updates and silently let users mismatch fields across extras.

Hoist to payments_py.x402.types (already canonical for shared x402
types) and import from there in both middleware modules. The Starlette
Request type is referenced via TYPE_CHECKING + forward string so
types.py stays importable in installations without the
[fastapi]/[langsmith] extras — at runtime the dataclass field
annotation is never evaluated.

The LangChain decorator's CreditsCallable is intentionally NOT
consolidated: it uses `Callable[..., int]` with a {"args", "result"}
dict shape, whereas the two middlewares share
`Callable[[Request], Union[int, Awaitable[int]]]`. Different signatures
for different lifecycle stages — wouldn't be safe to share.

Both submodules still re-export RouteConfig + CreditsCallable as
top-level names, so all existing user imports
(`from payments_py.x402.fastapi import RouteConfig`,
`from payments_py.langsmith import RouteConfig`) keep working.

523 of 523 unit tests green. No behavior change.
